Move tradeOrders to orders module
