+++ src/modules/orders/index.ts
@@ -1,1 +1,3 @@
 export * from './orders';
+export * from './tradeOrders';
+export * from './types';
