+++ src/modules/orders/tradeOrders.ts
@@ -3,9 +3,7 @@ import { useState, useEffect, useContext } from 'react';
 import { FirebaseContext } from '../firebase';
 import { Configuration } from '../stock';
 import { Profile } from '../../components/forms/ProfileInfoForm';
-
-export const ORDER_STATUSES = ['pending', 'approved', 'scheduled', 'completed'];
-export type OrderStatus = 'pending' | 'approved' | 'scheduled' | 'completed';
+import { OrderStatus } from './types';
 
 interface TradeOrderMetaData {
   creationTimestamp: Date;
