+++ src/modules/orders/types.ts
@@ -1,0 +1,13 @@
+export const ORDER_STATUSES = [
+  'pending',
+  'approved',
+  'scheduled',
+  'completed',
+  'cancelled',
+];
+export type OrderStatus =
+  | 'pending'
+  | 'approved'
+  | 'scheduled'
+  | 'completed'
+  | 'cancelled';
