--- src/modules/tradeOrders/index.ts
@@ -1,1 +1,0 @@
-export * from './tradeOrders';
